Add a hackish way of building and creating a separate pgcrypto
Based upon the work in johto postgresql repo, steal the pgcrypto
extension, build it quick and dirty, rename it and build into a debian
package. I would ideally like to build this upon the existing
contrib-9.3 packages and simply replace them, but the source built upon
has diverged form this and the work included in making the existing
packages work is simply not worth it at this point. This is a a bit of a
hack.