Accept messages received on non-Primary keys
We have to accept messages to any keys which are still in the database
to provide a reasonable way to change/upgrade keys with no downtime.
Note that we also assume that any key still in the table will not have
been revoked (but there's no "official" way to revoke a key as of yet).