When receiving a Message, set Messages.Datestamp to the "creation_time" extracted from the PGP signature.
This is more useful, as we can then both see when the message was created/sent from the sender (Messages.Datestamp),
and when we received it (Messages.Delivered). Previously, Messages.Datestamp was always equal to Messages.Delivered
for messages received, and only differed for messages sent. This improvement is possible thanks to the new
value "creation_time" returned by pgcrypto pgp_pub_signatures().