Fix bug in the handling of MDC when extracting signature information
I had originally thought that it's OK not to care about the MDC if we
skip over the data, but that appears to not be exactly true.  There's
some corner case where ignoring the MDC in a "stream" context would
cause the compressed data packet to be only partially read, resulting in
weird errors about packet headers.  Fix by fully handling MDC, just like
pgp-encrypt.c does.

Additionally, update to the version of the armor header patch which got
pushed upstream in commit 32984d8fc3dbb90a3fafb69fece0134f1ea790f9.  The
C-function interface changed slightly, but we can provide a trivial SQL
function to provide the same functionality.

Bump the version of the deb package to 1.3-1 and the PostgreSQL
extension to 1.3 due to the interface change in the armor header
functionality.  ALTER EXTENSION pgcrypto_openpgp UPDATE TO '1.3'; will
be necessary after upgrading the package.  Unfortunately it's possible
to see the error 'could not find function "pgp_armor_header_w"' between
installing the .deb and running the ALTER EXTENSION.  Caveat emptor.